Revert "Update of Ankr project and refactoring of OnX one"
