--- projects/config/ankr/avalanche/tokenAddresses.js
@@ -1,5 +1,0 @@
-const tokenAddresses = {
-  aAVAXb: "0x6C6f910A79639dcC94b4feEF59Ff507c2E843929",
-};
-
-module.exports = tokenAddresses;
