--- projects/config/ankr/ethereum/tokenAddresses.js
@@ -1,8 +1,0 @@
-const tokenAddresses = {
-  aETHc: "0xE95A203B1a91a908F9B9CE46459d101078c2c3cb",
-  aMATICb: "0x99534ef705df1fff4e4bd7bbaaf9b0dff038ebfe",
-  aDOTb: "0x5cc56c266143f29a5054b9ae07f3ac3513a7965e",
-  aKSMb: "0x84da8e731172827fcb233b911678e2a82e27baf2",
-};
-
-module.exports = tokenAddresses;
