+++ projects/config/onx/ethereum/abis/slp-vault-abi.json
