+++ projects/config/onx/ethereum/farmTvl.js
@@ -1,6 +1,7 @@
 const BigNumber = require('bignumber.js');
-const { ZERO, getTotalSupplyOf, getBalanceOf } = require('../../../helper/ankr/utils');;
-const ERC20Abi = require('../../../helper/ankr/abis/ERC20.json');
+const { farmContracts } = require('./farms');
+const { ZERO, getTotalSupplyOf, getBalanceOf } = require('../utils');
+const ERC20Abi = require('./abis/ERC20.json');
 const web3 = require('../../web3.js');
 
 const getSymbolPrice = (
